marlin: sched: rt: Avoid CPUs running top apps if possible
If an RT task was running on a CPU some time back and a top-app happen
to be running later, then when the RT task wakes up, we will blindly
wake up the RT task on its previous CPU thus preventing the top-app from
running until the RT task sleeps. RT tasks can run on any available CPU
that's lower in priority than itself so we should look at all possible
options under this condition. Incase no option is available, it will
fallback to preempting the top-app as before.

Tests:
* Improvement in UiBench GLTextureView seen which gets us to the N levels.
  Other benchmarks show good or better performance.
* No energy increases seen with UiBench GLTextureView and Camera preview.

bug: 62806392

Change-Id: Ifea9157c23decc34f64ce8ab87df8ac4d1bcb3c3
Signed-off-by: Joel Fernandes <joelaf@google.com>
Signed-off-by: joshuous <joshuous@gmail.com>
Signed-off-by: Dede Dindin Qudsy <xtrymind@gmail.com>